refactor: move enotdir to its own module
