+++ rust/src/tests/errors.rs
@@ -30,123 +30,7 @@ use crate::{
 
 mod eloop;
 mod enoent;
-
-crate::test_case! {enotdir => [Regular, Fifo, Block, Char, Socket]}
-fn enotdir(ctx: &mut TestContext, ft: FileType) {
-    /// Asserts that it returns ENOTDIR if a component of the path prefix is not a directory.
-    fn assert_enotdir<T: Debug, F>(ctx: &mut TestContext, ft: &FileType, f: F)
-    where
-        F: Fn(&Path) -> nix::Result<T>,
-    {
-        let base_path = ctx.create(ft.clone()).unwrap();
-        let path = base_path.join("previous_not_dir");
-
-        assert_eq!(f(&path).unwrap_err(), Errno::ENOTDIR);
-    }
-
-    /// Asserts that it returns ENOTDIR if a component of either path prefix is not a directory.
-    fn assert_enotdir_two_params<T: Debug, F>(ctx: &mut TestContext, ft: &FileType, f: F)
-    where
-        F: Fn(&Path, &Path) -> nix::Result<T>,
-    {
-        let base_path = ctx.create(ft.clone()).unwrap();
-        let path = base_path.join("previous_not_dir");
-        let new_path = ctx.base_path().join("not_existent");
-
-        assert_eq!(f(&path, &new_path).unwrap_err(), Errno::ENOTDIR);
-
-        let file = ctx.create(ft.clone()).unwrap();
-
-        assert_eq!(f(&file, &path).unwrap_err(), Errno::ENOTDIR);
-    }
-
-    /// Asserts that it returns ENOTDIR when the 'from' argument is a directory, but 'to' is not a directory.
-    fn assert_enotdir_from_to<T: Debug, F>(ctx: &mut TestContext, ft: &FileType, f: F)
-    where
-        F: Fn(&Path, &Path) -> nix::Result<T>,
-    {
-        let path = ctx.create(ft.clone()).unwrap();
-        let dir = ctx.create(FileType::Dir).unwrap();
-
-        assert_eq!(f(&dir, &path).unwrap_err(), Errno::ENOTDIR);
-    }
-
-    #[cfg(any(
-        target_os = "openbsd",
-        target_os = "netbsd",
-        target_os = "freebsd",
-        target_os = "dragonfly",
-        target_os = "macos",
-        target_os = "ios"
-    ))]
-    {
-        // chflags/01.t
-        assert_enotdir(ctx, &ft, |p| chflags(p, FileFlag::empty()));
-    }
-    // chmod/01.t
-    assert_enotdir(ctx, &ft, |p| chmod(p, Mode::empty()));
-
-    #[cfg(any(target_os = "netbsd", target_os = "freebsd", target_os = "dragonfly"))]
-    {
-        assert_enotdir(ctx, &ft, |p| lchmod(p, Mode::empty()));
-    }
-    // chown/01.t
-    assert_enotdir(ctx, &ft, |p| {
-        chown(
-            p,
-            Some(User::from_name("nobody").unwrap().unwrap().uid),
-            None,
-        )
-    });
-    assert_enotdir(ctx, &ft, |p| {
-        lchown(
-            p,
-            Some(User::from_name("nobody").unwrap().unwrap().uid),
-            None,
-        )
-    });
-    // link/01.t
-    assert_enotdir_two_params(ctx, &ft, |p1, p2| link(p1, p2));
-    // mkdir/01.t
-    assert_enotdir(ctx, &ft, |p| mkdir(p, Mode::empty()));
-    // mkfifo/01.t
-    assert_enotdir(ctx, &ft, |p| mkfifo(p, Mode::empty()));
-    // mknod/01.t
-    assert_enotdir(ctx, &ft, |p| mknod(p, SFlag::S_IFIFO, Mode::empty(), 0));
-    // open/01.t
-    assert_enotdir(ctx, &ft, |p| open(p, OFlag::O_RDONLY, Mode::empty()));
-    assert_enotdir(ctx, &ft, |p| {
-        open(p, OFlag::O_CREAT, Mode::from_bits_truncate(0o644))
-    });
-    // rename/12.t
-    assert_enotdir_two_params(ctx, &ft, |p1, p2| rename(p1, p2));
-    // rename/13.t
-    assert_enotdir_from_to(ctx, &ft, |from, to| rename(from, to));
-    // rmdir/01.t
-    assert_enotdir(ctx, &ft, rmdir);
-    // symlink/01.t
-    assert_enotdir(ctx, &ft, |p| symlink(Path::new("test"), p));
-    // (f)truncate/01.t
-    assert_enotdir(ctx, &ft, |p| truncate(p, 0));
-    // unlink/01.t
-    assert_enotdir(ctx, &ft, |p| unlink(p));
-}
-
-crate::test_case! {enotdir_privileged, root => [Regular, Fifo, Block, Char, Socket]}
-fn enotdir_privileged(ctx: &mut TestContext, ft: FileType) {
-    let base_path = ctx.create(ft.clone()).unwrap();
-    let path = base_path.join("previous_not_dir");
-
-    // mknod/01.t
-    assert_eq!(
-        mknod(&path, SFlag::S_IFCHR, Mode::empty(), 0).unwrap_err(),
-        Errno::ENOTDIR
-    );
-    assert_eq!(
-        mknod(&path, SFlag::S_IFBLK, Mode::empty(), 0).unwrap_err(),
-        Errno::ENOTDIR
-    );
-}
+mod enotdir;
 
 crate::test_case! {eacces, serialized, root}
 fn eacces(ctx: &mut SerializedTestContext) {
